Significant updates
Clause Input is Done
Target Input based on a list of types is done.
