+++ OAT.Benchmarks/ExpressionBenchmarks.cs
@@ -1,6 +1,7 @@
 ﻿using BenchmarkDotNet.Attributes;
 using Microsoft.CST.OAT;
 using Microsoft.CST.OAT.Tests;
+using Microsoft.CST.OAT.VehicleDemo;
 using System.Collections.Generic;
 using static Microsoft.CST.OAT.Tests.VehicleDemo;
