+++ OAT.Blazor/ComponentCollapsedState.cs
@@ -1,0 +1,8 @@
+namespace Microsoft.CST.OAT.Blazor.Components
+{
+    public enum ComponentCollapsedState
+    {
+        Expanded,
+        Collapsed
+    }
+}
