+++ OAT.Blazor/OAT.Blazor.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
@@ -15,8 +15,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\OAT.Tests\OAT.Tests.csproj" />
     <ProjectReference Include="..\OAT\OAT.csproj" />
+    <ProjectReference Include="..\VehicleDemo\VehicleDemo.csproj" />
   </ItemGroup>
 
   <ItemGroup>
