+++ OAT.Blazor/Pages/Author.razor
@@ -1,0 +1,7 @@
+﻿@page "/author"
+
+<h3>Author</h3>
+This is the page for authoring rules.
+@code {
+
+}
