+++ OAT.Blazor/Pages/Demo.razor
@@ -1,0 +1,7 @@
+﻿@page "/demo"
+
+<h3>Demo</h3>
+This is the sandbox page but with the demo rules loaded?
+@code {
+
+}
