+++ OAT.Blazor/Pages/Index.razor
@@ -1,8 +1,6 @@
 ﻿@page "/"
-@using Microsoft.CST.OAT.Blazor.Components;
-@using Microsoft.CST.OAT.Tests;
 
-<RuleInput Rule="rule" />
+<RuleInput Rule="rule" CollapsedState="@ComponentCollapsedState.Expanded"/>
 
 @code{
     Rule rule = new Rule("Default Rule");
