+++ OAT.Blazor/Pages/Sandbox.razor
@@ -1,0 +1,7 @@
+﻿@page "/sandbox"
+
+<h3>Sandbox</h3>
+This is the sandbox for creating objects and testing them against rules.
+@code {
+
+}
