+++ OAT.Blazor/Shared/NavMenu.razor
@@ -13,13 +13,18 @@
             </NavLink>
         </li>
         <li class="nav-item px-3">
-            <NavLink class="nav-link" href="counter">
-                <span class="oi oi-plus" aria-hidden="true"></span> Counter
+            <NavLink class="nav-link" href="demo">
+                <span class="oi oi-plus" aria-hidden="true"></span> Demo
             </NavLink>
         </li>
         <li class="nav-item px-3">
-            <NavLink class="nav-link" href="fetchdata">
-                <span class="oi oi-list-rich" aria-hidden="true"></span> Fetch data
+            <NavLink class="nav-link" href="author">
+                <span class="oi oi-list-rich" aria-hidden="true"></span> Author
+            </NavLink>
+        </li>
+        <li class="nav-item px-3">
+            <NavLink class="nav-link" href="sandbox">
+                <span class="oi oi-list-rich" aria-hidden="true"></span> Sandbox
             </NavLink>
         </li>
     </ul>
