+++ OAT.Blazor/_Imports.razor
@@ -5,5 +5,6 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using OAT.Blazor
-@using OAT.Blazor.Shared
+@using Microsoft.CST.OAT.Blazor.Shared;
+@using Microsoft.CST.OAT.Blazor.Components;
+@using Microsoft.CST.OAT.Utils;
