+++ OAT.Tests/OAT.Tests.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net48;netcoreapp3.1</TargetFrameworks>
+    <TargetFrameworks>net48;netcoreapp3.1;</TargetFrameworks>
 
     <IsPackable>false</IsPackable>
 
@@ -23,6 +23,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\OAT\OAT.csproj" />
+    <ProjectReference Include="..\VehicleDemo\VehicleDemo.csproj" />
   </ItemGroup>
 
   <ItemGroup>
