+++ OAT.sln
@@ -14,7 +14,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OAT.Benchmarks", "OAT.Benchmarks\OAT.Benchmarks.csproj", "{F7BF4714-746B-4D02-9BA0-FFC0B2FF28C1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OAT.Blazor", "OAT.Blazor\OAT.Blazor.csproj", "{279356E4-F42C-4213-B6FD-902765C16384}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OAT.Blazor", "OAT.Blazor\OAT.Blazor.csproj", "{279356E4-F42C-4213-B6FD-902765C16384}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VehicleDemo", "VehicleDemo\VehicleDemo.csproj", "{8FB5D359-73F0-45D1-BAC3-DC54242E54C6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -38,6 +40,10 @@ Global
 		{279356E4-F42C-4213-B6FD-902765C16384}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{279356E4-F42C-4213-B6FD-902765C16384}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{279356E4-F42C-4213-B6FD-902765C16384}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8FB5D359-73F0-45D1-BAC3-DC54242E54C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8FB5D359-73F0-45D1-BAC3-DC54242E54C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8FB5D359-73F0-45D1-BAC3-DC54242E54C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8FB5D359-73F0-45D1-BAC3-DC54242E54C6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
